Move the low-level allocator implementations to a header file.
